Decide on mcaWritingStyleFormatter to encapsulate some "utils"
