+++ src/components/BeaconsTable.tsx
@@ -25,9 +25,8 @@ import React, {
 } from "react";
 import { IBeacon } from "../entities/IBeacon";
 import { IBeaconsGateway } from "../gateways/IBeaconsGateway";
-import { formatDate } from "../useCases/formatDate";
-import { formatUses } from "../useCases/formatUses";
-import { titleCase } from "../utils";
+import { formatDate, titleCase } from "../useCases/mcaWritingStyleFormatter";
+import { IUse } from "../entities/IUse";
 
 interface IBeaconsTableProps {
   beaconsGateway: IBeaconsGateway;
@@ -163,4 +162,16 @@ export const BeaconsTable: FunctionComponent<IBeaconsTableProps> = ({
       }}
     />
   );
+};
+
+export const formatUses = (uses: IUse[]): string =>
+  uses.reduce((formattedUses, use, index, uses) => {
+    if (index === uses.length - 1) return formattedUses + formatUse(use);
+    return formattedUses + formatUse(use) + ", ";
+  }, "");
+
+const formatUse = (use: IUse): string => {
+  const formattedActivity = titleCase(use.activity);
+  const formattedPurpose = use.purpose ? ` (${titleCase(use.purpose)})` : "";
+  return formattedActivity + formattedPurpose;
 };
