--- src/useCases/formatDate.ts
@@ -1,4 +1,0 @@
-export const formatDate = (date: Date): string => {
-  const [, month, day, year] = date.toDateString().split(" ");
-  return `${parseInt(day)} ${month} ${year.slice(2)}`;
-};
