+++ src/useCases/mcaWritingStyleFormatter.ts
@@ -1,3 +1,8 @@
+export const formatDate = (date: Date): string => {
+  const [, month, day, year] = date.toDateString().split(" ");
+  return `${parseInt(day)} ${month} ${year.slice(2)}`;
+};
+
 export const titleCase = (text: string): string => {
   return text
     .replace("_", " ")
